Moved Discord module next to Discord platform
