+++ lib/src/platform/discord/discord_module.dart
